+++ src/main/java/message_send/application_business_rule/MessageOutputBoundary.java
@@ -1,4 +1,4 @@
-package message_sebd.application_business_rule;
+package message_send.application_business_rule;
 
 public interface MessageOutputBoundary {
     MessageResponseModel prepareSuccessView(MessageResponseModel response);
