+++ src/main/java/message_send/interface_adaptor/MessageCreationFailed.java
@@ -1,4 +1,4 @@
-package message_sebd.interface_adaptor;
+package message_send.interface_adaptor;
 
 public class MessageCreationFailed extends RuntimeException {
     public MessageCreationFailed(String error) {
